Fix tests

+++ java/test/good/14.cc
@@ -9,24 +9,24 @@
 // This program approximates pi and then checks that the approximation is within
 // an acceptable bound
 
-double piApprox(int iterations)
+double piApprox(double iterations)
 {
     double pi = 0.0;
-    double factor = 1;
-    int i = 2;
+    double factor = 1.0;
+    double i = 2.0;
     while (i < iterations)
     {
         pi = pi + factor / (i - 1.0);
-        i = i + 2;
-        factor = factor * (0 - 1);
+        i = i + 2.0;
+        factor = factor * (0.0 - 1.0);
     }
-    return 4 * pi;
+    return 4.0 * pi;
 }
 
 int main()
 {
     printString("How many iterations: ");
-    int iterations = readInt();
+    double iterations = readDouble();
     double pi = piApprox(iterations);
     // printDouble(pi);
     if (pi < 3.20 && pi > 3.10)
